Modularize levm and revm functionality
