+++ crates/vm/Cargo.toml
@@ -6,9 +6,10 @@ edition = "2021"
 [dependencies]
 ethrex-core = { path = "../common", default-features = false }
 ethrex-storage = { path = "../storage/store", default-features = false }
-ethrex-levm = { path = "./levm", optional = true }
+ethrex-levm = { path = "./levm", optional = false }
 ethrex-trie = { path = "../storage/trie", default-features = false }
 ethrex-rlp = { path = "../common/rlp", default-features = false }
+
 revm = { version = "14.0.3", features = [
   "serde",
   "std",
@@ -39,7 +40,6 @@ path = "./vm.rs"
 
 [features]
 default = ["c-kzg", "blst"]
-levm = ["default", "ethrex-levm"]
 l2 = []
 c-kzg = ["revm/c-kzg"]
 blst = ["revm/blst"]
